includes a scribe for just faster whisperer
